plugins/agentops-meta-orchestrator/skills/agentops-orchestrator/scripts/requirements.txt
# AgentOps Meta-Orchestrator - Neo4j Integration Dependencies
# Install: pip install -r requirements.txt

# Neo4j Python Driver
# Used by: neo4j_connector.py, migrate_to_neo4j.py, pattern_storage.py
neo4j>=5.14.0,<6.0.0

# YAML Processing
# Used by: pattern_storage.py, pattern_matcher.py, migrate_to_neo4j.py
PyYAML>=6.0.1

# Optional: For development and testing
pytest>=7.4.0  # Unit testing
pytest-cov>=4.1.0  # Code coverage
black>=23.0.0  # Code formatting
flake8>=6.1.0  # Linting
mypy>=1.6.0  # Type checking
